refator: rename few .js files
