+++ index.js
@@ -1,20 +1,27 @@
 import express from "express";
 import mongoose from "mongoose";
-import router from "./routes/cars-router.js";
 import * as dotenv from "dotenv";
+import carsRouter from "./routes/cars.js";
 dotenv.config();
 
+// App
 const app = express();
+
+// Middlewares
 app.use(express.json());
-app.use("/api", router);
+
+// Routes
+app.use("/api", carsRouter);
 
 async function startApp() {
   try {
     mongoose.set("strictQuery", false);
-    await mongoose.connect(process.env.DB_URL, {
+    const connect = await mongoose.connect(process.env.DB_URL, {
       useNewUrlParser: true,
       useUnifiedTopology: true,
     });
+    console.log(`DB host: ${connect.connection.host}`);
+
     app.listen(process.env.PORT, () => console.log("server started"));
   } catch (e) {
     console.warn(e.message);
